Refactored route structure, and component names
