+++ src/global.d.ts
@@ -2,3 +2,5 @@ declare module "*.jpeg" {
   const value: string;
   export = value;
 }
+
+declare module 'react-router-dom'
